+++ src/tor_proxy.rs
@@ -13,10 +13,10 @@ use k8s_openapi::{
             HorizontalPodAutoscalerSpec, MetricSpec,
         },
         core::v1::{
-            Affinity, Capabilities, ConfigMap, ConfigMapVolumeSource, Container, ContainerPort,
-            ExecAction, KeyToPath, LocalObjectReference, PodSecurityContext, PodSpec,
-            PodTemplateSpec, Probe, SecurityContext, Service, ServicePort, ServiceSpec, Toleration,
-            TopologySpreadConstraint, Volume, VolumeMount,
+            Affinity, Capabilities, ConfigMap, ConfigMapVolumeSource, Container, ExecAction,
+            KeyToPath, LocalObjectReference, PodSecurityContext, PodSpec, PodTemplateSpec, Probe,
+            SecurityContext, Service, ServicePort, ServiceSpec, Toleration,
+            TopologySpreadConstraint, Volume,
         },
     },
     apiextensions_apiserver::pkg::apis::apiextensions::v1::CustomResourceDefinition,
@@ -38,7 +38,7 @@ use crate::{
     kubernetes::{
         self, error_policy, pod_security_context, Annotations, Api, ConditionsExt,
         Container as KubernetesContainer, Labels, Object, Resource as KubernetesResource,
-        ResourceName, SelectorLabels, Torrc as KubernetesTorrc,
+        ResourceName, SelectorLabels, Torrc as KubernetesTorrc, Volume as KubernetesVolume,
     },
     metrics::Metrics,
     tor::Torrc,
@@ -142,6 +142,9 @@ pub struct TorProxySpecDeployment {
 
     /// TopologySpreadConstraints describes how a group of pods ought to spread across topology domains. Scheduler will schedule pods in a way which abides by the constraints. All topologySpreadConstraints are ANDed.
     pub topology_spread_constraints: Option<Vec<TopologySpreadConstraint>>,
+
+    /// List of volumes that can be mounted by containers belonging to the pod. More info: https://kubernetes.io/docs/concepts/storage/volumes
+    pub volumes: Option<BTreeMap<String, KubernetesVolume>>,
 }
 
 fn default_deployment_replicas() -> i32 {
@@ -376,6 +379,16 @@ impl TorProxy {
             .as_ref()
             .and_then(|f| f.topology_spread_constraints.as_ref())
             .cloned()
+    }
+
+    #[must_use]
+    pub fn deployment_volumes(&self) -> BTreeMap<String, KubernetesVolume> {
+        self.spec
+            .deployment
+            .as_ref()
+            .and_then(|f| f.volumes.as_ref())
+            .cloned()
+            .unwrap_or_default()
     }
 
     #[must_use]
@@ -907,20 +920,7 @@ fn generate_deployment(
                     security_context: Some(object.deployment_security_context()),
                     tolerations: object.deployment_tolerations(),
                     topology_spread_constraints: object.deployment_topology_spread_constraints(),
-                    volumes: Some(vec![Volume {
-                        name: "etc-configs".into(),
-                        config_map: Some(ConfigMapVolumeSource {
-                            default_mode: Some(0o400),
-                            items: Some(vec![KeyToPath {
-                                key: "torrc".into(),
-                                mode: Some(0o400),
-                                path: "torrc".into(),
-                            }]),
-                            name: Some(object.config_map_name().into()),
-                            optional: Some(false),
-                        }),
-                        ..Default::default()
-                    }]),
+                    volumes: Some(generate_deployment_volumes(object)),
                     ..Default::default()
                 }),
             },
@@ -931,11 +931,7 @@ fn generate_deployment(
 }
 
 fn generate_deployment_containers(object: &TorProxy, config: &Config) -> Vec<Container> {
-    let mut containers = object
-        .deployment_containers()
-        .into_iter()
-        .map(|(k, v)| (k.clone(), v.to_container(k)))
-        .collect::<BTreeMap<_, _>>();
+    let mut containers = object.deployment_containers();
 
     {
         let container = containers.entry("tor".to_string()).or_default();
@@ -971,21 +967,21 @@ fn generate_deployment_containers(object: &TorProxy, config: &Config) -> Vec<Con
             timeout_seconds: Some(1),
             ..Default::default()
         });
-        container.name = "tor".into();
-        container.ports = Some(vec![
-            ContainerPort {
-                container_port: 1080,
-                name: Some("http-tunnel".to_string()),
-                protocol: Some("TCP".to_string()),
-                ..Default::default()
-            },
-            ContainerPort {
-                container_port: 9050,
-                name: Some("socks".to_string()),
-                protocol: Some("TCP".to_string()),
-                ..Default::default()
-            },
-        ]);
+
+        let ports = container.ports.get_or_insert_with(Default::default);
+
+        {
+            let port = ports.entry("http-tunnel".to_string()).or_default();
+            port.container_port = 1080;
+            port.protocol = Some("TCP".to_string());
+        }
+
+        {
+            let port = ports.entry("socks".to_string()).or_default();
+            port.container_port = 9050;
+            port.protocol = Some("TCP".to_string());
+        }
+
         container.readiness_probe = Some(Probe {
             exec: Some(ExecAction {
                 command: Some(vec![
@@ -1000,27 +996,57 @@ fn generate_deployment_containers(object: &TorProxy, config: &Config) -> Vec<Con
             timeout_seconds: Some(1),
             ..Default::default()
         });
-        container.volume_mounts = Some(vec![VolumeMount {
-            mount_path: "/etc/configs".into(),
-            name: "etc-configs".into(),
-            read_only: Some(true),
-            ..Default::default()
-        }]);
+
+        let volume_mounts = container.volume_mounts.get_or_insert_with(Default::default);
+
+        {
+            let volume_mount = volume_mounts.entry("etc-configs".to_string()).or_default();
+            volume_mount.mount_path = "/etc/configs".into();
+            volume_mount.read_only = Some(true);
+        }
     }
 
-    for container in containers.values_mut() {
-        container.security_context = Some(SecurityContext {
-            capabilities: Some(Capabilities {
-                drop: Some(vec!["ALL".to_string()]),
+    let mut containers = containers
+        .into_iter()
+        .map(|(name, value)| value.into_container(name))
+        .map(|mut container| {
+            container.security_context = Some(SecurityContext {
+                capabilities: Some(Capabilities {
+                    drop: Some(vec!["ALL".to_string()]),
+                    ..Default::default()
+                }),
                 ..Default::default()
-            }),
-            ..Default::default()
+            });
+            container
+        })
+        .collect::<Vec<_>>();
+    containers.sort_by(|a, b| a.name.cmp(&b.name));
+    containers
+}
+
+fn generate_deployment_volumes(object: &TorProxy) -> Vec<Volume> {
+    let mut volumes = object.deployment_volumes();
+
+    {
+        let volume = volumes.entry("etc-configs".to_string()).or_default();
+        volume.config_map = Some(ConfigMapVolumeSource {
+            default_mode: Some(0o400),
+            items: Some(vec![KeyToPath {
+                key: "torrc".into(),
+                mode: Some(0o400),
+                path: "torrc".into(),
+            }]),
+            name: Some(object.config_map_name().into()),
+            optional: Some(false),
         });
     }
 
-    let mut containers = containers.into_values().collect::<Vec<_>>();
-    containers.sort_by(|a, b| a.name.cmp(&b.name));
-    containers
+    let mut volumes = volumes
+        .into_iter()
+        .map(|(name, value)| value.into_volume(name))
+        .collect::<Vec<_>>();
+    volumes.sort_by(|a, b| a.name.cmp(&b.name));
+    volumes
 }
 
 fn generate_horizontal_pod_autoscaler(